ComponentDialog: Unify wiht ComponentDialogBase as currently nobody benefits from segregating these abstractions
